some changes in run-test.sh

+++ tests/run-test.sh
@@ -6,7 +6,7 @@ BACKEND=$2
 SCENARIO_FILE=$(basename "$SCENARIO")
 SCENARIO_NAME="${SCENARIO_FILE%.*}-${BACKEND}"
 
-STORAGE=`mktemp -d ${SCENARIO_NAME}-XXXXXXXX`
+STORAGE=`mktemp -d storage-${SCENARIO_NAME}-XXXXXXXX`
 
 timeout 60 ./mraft-py-test na+sm \
     -n 3                         \
@@ -17,6 +17,8 @@ timeout 60 ./mraft-py-test na+sm \
     -t $STORAGE
 RET=$?
 
-rm -rf $STORAGE
+if [ "$RET" -eq "0" ]; then
+    rm -rf $STORAGE
+fi
 
 exit $RET
